Separasão do projeto em arquivos (beta)
