+++ funcCadastrar.h
@@ -1,0 +1,12 @@
+
+#ifndef FUNCOES_H
+#define FUNCOESCAD_H
+
+#include "estruturas.h"
+
+void mostrarLista(Lista* lista);
+void Cadastrar(Lista* lista);
+void tratamento(char *entrada,char *novaEntrada);
+
+#endif
+
